server/pyproject.toml
[project]
name = "gtfs-rt-rater"
version = "0.1.0"
description = "Chalice API for the GTFS-RT Feed Quality Rater"
requires-python = ">=3.12,<3.13"
license = "MIT"
dependencies = [
    "requests~=2.32.5",
    "boto3~=1.39.0",
    "chalice-spec~=0.7.0",
    "pydantic~=2.11.7",
    "apispec~=6.8.2",
]

[dependency-groups]
dev = [
    "chalice==1.32.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.14.7",
    "pytest~=9.0.2",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
reportMissingImports = true
reportMissingTypeStubs = false
useLibraryCodeForTypes = false
